+++ saq/collectors/email/journal.py
@@ -182,6 +182,7 @@ class JournalEmailCollector(Collector):
                 email_observable.add_directive(DIRECTIVE_NO_SCAN)
                 email_observable.add_directive(DIRECTIVE_ORIGINAL_EMAIL)
                 email_observable.add_directive(DIRECTIVE_ARCHIVE)
+                email_observable.add_yara_meta("type", "email")
 
             yield Submission(root, key=file_name)
 
@@ -250,6 +251,7 @@ class JournalEmailCollector(Collector):
                 email_observable.add_directive(DIRECTIVE_NO_SCAN)
                 email_observable.add_directive(DIRECTIVE_ORIGINAL_EMAIL)
                 email_observable.add_directive(DIRECTIVE_ARCHIVE)
+                email_observable.add_yara_meta("type", "email")
 
             yield Submission(root, key=object_key)